[change] fixed scenegraph parser calls in viewport override constructor, moved model parsing code from scenegraph parser to model parser
